Fixed rating not accepting 0.5 values

+++ app/models/SchoolRating.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, ForeignKey, Integer, String
+from sqlalchemy import Column, Float, ForeignKey, Integer, String
 from sqlalchemy.orm import Mapped
 from app.db.db import db
 
@@ -6,7 +6,7 @@ class SchoolRating(db.Model):
     __tablename__ = "school_rating"
 
     id: Mapped[int] = Column(Integer, primary_key=True)
-    rating: Mapped[int] = Column(Integer)
+    rating: Mapped[int] = Column(Float)
     comment: Mapped[str] = Column(String(256))
     
     user_id: Mapped[int] = Column(Integer, ForeignKey('user.id'))
